去除编译告警
git-svn-id: http://192.168.20.11/sequoiadb/trunk@22777 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/rtn/rtnContext.cpp
@@ -2967,7 +2967,7 @@ namespace engine
       EMPTY_CONTEXT_MAP::iterator iter ;
       coordSubContext *pSubContext = NULL ;
 
-      if ( (UINT32)pReply->header.opCode != MSG_BS_GETMORE_RES ||
+      if ( pReply->header.opCode != MSG_BS_GETMORE_RES ||
            (UINT32)pReply->header.messageLength < sizeof( MsgOpReply ) )
       {
          rc = SDB_UNKNOWN_MESSAGE ;
